Backend/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
asttokens @ file:///croot/asttokens_1743630435401/work
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
coloredlogs==15.0.1
comm @ file:///croot/comm_1709322850197/work
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
distro==1.9.0
doctr==1.9.0
etelemetry==0.3.1
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.7.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
flatbuffers==25.2.10
fonttools==4.59.2
frontend==0.0.3
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.31.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.30.2
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
ipykernel @ file:///croot/ipykernel_1737660677549/work
ipython @ file:///croot/ipython_1734548052611/work
isodate==0.7.2
itsdangerous==2.2.0
jedi @ file:///croot/jedi_1733987392413/work
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1751991368470/work
kiwisolver==1.4.9
langdetect==1.0.9
langextract==1.0.8
looseversion==1.3.0
lxml==6.0.1
magika==0.6.2
markdownify==1.2.0
markitdown==0.1.3
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
ml_collections==1.1.0
ml_dtypes==0.5.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnx==1.19.0
onnxruntime==1.22.1
openai==1.102.0
opencv-python==4.12.0.88
packaging @ file:///croot/packaging_1753775359799/work
pandas==2.3.1
parso @ file:///croot/parso_1733963305961/work
pathlib==1.0.1
pexpect @ file:///croot/pexpect_1750072593878/work
pillow==11.3.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
polars==1.33.1
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
Pygments @ file:///croot/pygments_1744664109463/work
PyMuPDF==1.26.4
pyparsing==3.2.3
pypdfium2==4.30.0
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-doctr==1.0.0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq @ file:///croot/pyzmq_1734687138743/work
RapidFuzz==3.14.1
rdflib==7.1.4
regex==2025.9.1
requests==2.32.5
rouge_score==0.1.2
rsa==4.9.1
safetensors==0.6.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.35.2
shapely==2.1.1
simplejson==3.20.1
six @ file:///croot/six_1744271502820/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1718227057033/work
traits==7.0.2
-e git+https://github.com/huggingface/transformers.git@8c60a7c385968f15969f08794139584d952e77f5#egg=transformers
triton==3.4.0
typing-inspection==0.4.1
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
ultralytics==8.3.198
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
wandb==0.21.3
wcwidth @ file:///croot/wcwidth_1750352883074/work
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
